Allow separate target nets in GreedyAC
Before, target networks depended on the output of the ensemble. This is
not the best idea when using an ensemble based on optimism, such as the
percentile critic ensemble, since we could artificially inflate Q
values. The idea is that the optimism should affect the actor update
but not the critic update.

This commit allows using a separate target value for each network in a
critic ensemble, rather than the (aggregated) output of the critic
itself. This is controlled by the `self.ensemble_targets` flag in
`GreedyAC`. Each update target is produced by a separate target network
in the critic ensemble. For example, for a critic ensemble of N
networks, each of the N networks produces an output, and each of these
outputs are used to construct the (separate) target value for the N
updates of the N corresponding networks in the critic ensemble.