refactor: clean up cfg prep for logging to sql
This change creates some separation between business logic and
operationalization of that logic with our particular data structures.

The `prep_cfg_...` function now directly takes a cfg, does whatever prep
work is needed, then does the dirty details of dealing with python dicts
to prepare the object.

In the process, I also moved over to creating a new copied dict instead
of modifying in place (e.g. see the `dict_u.drop` utility function).
While technically we'll lose a few nanoseconds of performance, freeing
ourselves of side-effects is likely worth it. It does not appear this
code is on the hotpath.