fix: add nullability type checks
Current buffer implementation has several possible codepaths that return
None. These `None`s aren't handled throughout, likely because
`self.data` is always defined by the time the downstream methods are
called. This is a bit of a footgun in the design that we should cover
more explicitly. For now, just marking appropriate methods as nullable
makes it easier to identify the refactoring opportunity down the line.