test: add ability to run opc tests in parallel
The major challenge was that we spin up a fresh OPC test server for
every test, which requires a (potentially) unique port for every test.
In order to guarantee each server has a unique port, we ask the OS
for an arbitrary free port (using special port `:0`).

However, both the client and the server need to agree on which port is
being used. In order to reduce duplication, added a new feature that
does this handshake automatically.