ci: assume async fixtures run in same loop as func
By default, async fixtures create their own event loop to run in,
which makes communicating between client and server difficult in the
test function. We can add an arg to the decorator to specify the desired
functionality (as was previously done), but `pytest-asyncio` will
still complaign in a warning message unless we set a global default
behavior.

I still like the explicitness of defining the behavior inline, so this
PR just adds the global default to silence the warning, while leaving
the `loop_scope='function'` parameter in the fixtures.