fix: avoid unnecessarily widening normalizer types
By annotating a function like:
```python
def f(a: int | str) -> int | str:
  return a
```
then calling the function with (e.g.)
```python
a: int = 0
c = f(a) # c has type int | str
```
then now `c` has an unnecessarily wide type, we know that `c` should be
an `int` because `a` was an `int`. We can avoid this type widening using
a variadic type:
```python
T = TypeVar('T', bound=int|str)
def f(a: T) -> T:
  return a
```
where now we communicate to the (type) compiler that the function `f`
returns the same type that it is given, type `T`, and that we can only
handle types `T` which are either an `int` or `str`.