feat: add dictionary hashing utility
The current hashing implementaiton is not robust to insertion order for
dictionaries, which means that reordering one of our configs will lead
to a different hash (even if the observable underlying config remains
the same).

This utility resolves that issue by ensuring a consistent ordering
by sorting the keys alphabetically.

Now that we are taking ownership of the hashing strategy, we can also
make it more feature complete to simplify the business logic throughout
the codebase. This is towards a larger initiative to separate business
logic from implementation detail. An example would be ignoring keys
while performing the hash (e.g. seed).