fix: resubscribe to events if client disconnects
The server will unsubscribe if it detects that a client disconnects in
order to prevent any memory leaks or linearly growing cost of subs with
a flaky client.

If for some reason the client disconnects and reconnects between
handshake events---thus the server never detected the disconnect---then
this will send a duplicate subscription event to the server. The server
already handles this gracefully and discards the duplicate event.