fix: fix typing of critic update closure
There were several issues in this codeblock, some of which I am kicking
down the road.

The `sum` built-in does not have appropriate overloads for summing over
tensors. This likely works, but not for the reasons that we would expect
which makes it brittle. Specifically, `sum` assumes that the initial
value is `0.` and luckly, `tensor + 0. = tensor` and the iteration
produces the desired result. The change to `vstack \comp torch.sum`
makes the expected behavior explicit, and also makes the type-checker
happy.

I don't understand the hooks strategy. However, they are preventing
the type-checker from doing its job. This particular hook appears unused
so I removed it.

The `batches` variable is defined in a loop body, the captured in a
function body within that loop. This type of variable capture is rife
with potential errors due to the fact the `batches` might have changed
by the time the `closure` function is actually called. I wasn't able to
easily fix this, so for now am silencing the linter as it appears this
is working as expected. A later refactor should likely confirm the
expected behavior here.