fix: narrow type annotation for q factory
The factory function for q learners has a fairly wide return type
annotation of BaseQ. This makes sense, except that only EnsembleQ
objects can actually be created at the moment. Consumers of this code
also strictly expect an EnsembleQ object to be returned.

For now, the easiest path forward is to tighten the type annotation to
match the implementation --- pushing the impetus of solving this issue
to whoever wants to add a non-EnsembleQ object to the factory.