chore: specify bins for FTA instead of `delta`
FTA has a requirement that the delta hyper divide the width of the
regions that the FTA is applied over. That is, if using an FTA over the
region `[l, u]`, then the hyperparameter `δ` of the FTA activation
should be set such that `(u - l) / δ = c₀` for some `c₀ > 0 ∈ ℤ`.

This complicated the constructor because these arguments were
implicitly interdependent. Thus, I've changed the constructor so that it
instead takes as input `c₀` and computed `δ`.