chore: warn when `head_` keys exist
The new interface for `create_base` ignores any activations,
initializations, etc. for keys in the config dict that start with
`head_`. This commit ensures warnings are raised when these keys exist
in the configuration and removes any assertions.

The `head_*` keys were rarely (never?) used. Further, the function name
`create_base` is slightly misleading in this case, since a base of a
network shouldn't have a head.