fix: remove OfflineAnytimeInteraction
A notable amount of the OfflineAnytimeInteraction code was broken,
e.g. the init function, the factory function, the integration of the
step function within main_utils.

The plumbing through the factory function does not currently enable us
to pass the needed arguments to the instantiation of the
OfflineAnytimeInteraction, requiring a notable rework to get the feature
working again. Due to the impending unification of online/offline data
injestion, getting this feature working seems irrelevant anyways.