refactor: rewrite normalizers using structured configs
Use the group dispatcher pattern to rewrite normalizers using structured
configs and enable deeper type-checking.

A consequence of this change is that nested config groups is a bit
harder (though not impossible) to maintain. The nested config group
pattern seemed unnecessarily complex for our usecase, so instead of
including this in the rewrite, I choose towards simplicity.
Unfortunately, this makes this refactor "impure" in the sense that part
of our API is changing alongside the rewrite.