feat: add utility for one-off structured configs
Some configs are not attached to a group, but still need to be
registered with hydra in order for schema validation to work. To make
this more semantic in consuming code, added a wrapper over the
`@dataclass` decorator.

The first user of this utility actually _is_ a hydra group, despite the
fact that we are accumulating a list of group members instead of a
discriminated union. Likely this is actually an abuse of hydra groups
instead of a valid use-case, but for the sake of minimizing breaking
changes during a refactor, we'll allow this utility to represent this
weird case.