refactor: switch BaseExploration to protocol
Generally, protocols make composition of parts much simpler than ABCs.
A protocol is not a `is a` relationship like class inheritence, instead
the protocol is a `acts like a` or `implements this interface` type of
relationship. In a vast majority of cases, this is really what we want.

The advantage to protocol over ABC is that it eliminates nominally typed
structures --- it doesn't matter if an exploration strategy strictly
inherits from BaseExploration, so long as it implements the interface
then we are happy.

If you are familiar with rust, this will feel pretty natural.