test: match pipeline test usage to intended use
Pipeline stages iteratively mutates the PipeFrame context object tag by
tag in order. By contrast, the tests pass the original PipeFrame into
each stage, overriding the output variable. This isn't actually
guaranteed to be semantically correct -- e.g.:
```python
a = f(a)
a = g(a)
```
doesn't guarantee that the results of `f` are actually represented in
the variable `a`. This works by happenstance currently because _all_
changes are captured in the mutated carry variable.

This change _also_ allows using much shorter variable names, which
heavily removes visual noise and makes the test easier to quickly grok.