feat: implement normalizer state_constructor
The normalizer can be configured either with a known min/max, or can be
told to estimate the min/max from data. It is also valid to configure
both, where the apriori min/max will be widened from observed data.

In order to realize the same behavior as before, we would need to mutate
the configs with a known min/max pulled from the environment before
instantiating the pipeline. This is a bit clunky and might be worth a
revisit at another time.

Completes: https://app.asana.com/0/1208460773823083/1208868219374007/f