feat: add split state construction
A very simple state constructor component that allows splitting the sc
pipeline and executing two children scs from the same root state.

This places no restrictions on what the left and right sc is --- it is
technically possible that these are also split scs --- and dealing with
that madness is up to the user.