fix: broadcast new missing info flag over mask
Currently, can successfully update the missing info flag if only a
single element of the mask is true. If multiple elements are true, then
this requires broadcasting over all selected rows. Unfortunately, pandas
doesn't support broadcasting over a mask-selector and requires the
new values to be the same shape.

This simply does that manually.