feat: add ability to mandate expected split passthrough behavior
The split transform current passes columns through based on the union
of the left and right subtransformation. If both left and right consume,
then the split consumes. If one of the left and right do not consume,
then the split does not consume.

This seems like expected default behavior, so this behavior is retained.
However, it also seems plausible that the user would want to strictly
specify "do not consume" and "do consume". If the user has expressed a
preference, then the preference overrides any default behavior.