feat: add ability to selectively enable and reorder pipeline stages
By default, the pipeline continues going through stages in a fixed
order and executes every stage. However, the invoker can selectively
choose to reorder or subselect pipeline stages to invoke.

This selection is done at call-time (on invocation) to allow us to have
a single pipeline instance with a single source-of-truth for state,
but the ability to reorganize that pipeline in arbitrary ways.

In a vast majority of cases, the order of stages should remain the same;
the expected use-case here is to enable/disable stages selectively.