refactor: use a single SC across all tags
Heavily simplifies the implementation and makes it easier to manage
the temporal state. Also enables the ability to have a single default
config used across all tags --- capturing the 95% case where every
tag uses the exact same set of state constructors.

This is also a necessary step to implementing state constructors that
are not strictly tied to a tag -- such as the countdown states within
the action period.