feat: add utility to get deeply nested values from dict
This currently **does not** support lists in the nested dict. That is
  `list_key[0].name`
is an invalid path. This limitation is due only to lack of consumer for
this more complex logic, so starting with the easy case until a consumer
appears.

Notably, the first consumer of this utility is the string interpolation
in configs. Since it is fundamentally hard to interpolate through a list
in the configs, there is no reason (yet) to support this.