refactor: replace hydra with interal config lib
The motivation is manyfold:
 - Hydra is complex and hard to extend properly
 - Hydra relies on global state and single-config assumptions
 - Hydra couples config directory layout with config shape
 - Hydra does not kindly support discriminated unions with structured
 configs

We are only using a small subset of Hydra's more advanced fetaures,
otherwise our configs are rather straightforward validated yaml files
using dataclasses to specify the schemas.

This commit implements _all_ of the features of hydra that we are using
with the sole exception of `interpolation` within a python dataclass.
This feature is a bit hard to reason about (schemas are holding onto
non-trivial default values) and so the current canonical approach will
be to specify interpolations within the yaml file instead.