refactor: simplify bound_checker implementation
Instead of having the bound-checking logic itself deal with messy
configurable values `float | None`, we first normalize the config to
have type `float`, then simplify the bound-checking logic to exclusively
do bound checking.

The original implementation is an example of the broader anti-pattern
of pushing `| None` types deep into the implementation instead of
normalizing the types earlier in the code.