refactor: pull empty dataframe handling out of stages and into pipeline
We had some inconsistent handling of empty dataframes in each stage.
Most stages were not handling this case at all, others were detecting
it differently, etc. Stages should assume that they receive a meaningful
and correct dataframe and should not have special case error handling.
Instead, it is the pipeline's job to make sure each stage is handed the
correct data --- and this allows consistency in how the data is treated.