feat: add config option to allow extra fields
We should use this option sparingly --- typically only for the top-level
schema and not for internal sub-parts of the schema. This option is
applied per-level.

This will allow us to reuse the same config instances (i.e. yaml files)
across different scripts, but force each script to specify its minimum
expected contract using the schemas. At the same time, this also
prevents us from having yaml files with configuration options that
nobody in the codebase uses (which happened before we started enforcing
schemas).

Note: for any deployed code, we should be careful of allowing extra
fields in the config. This is a common security attack vector. Through
explicit schemas and type-checking, we should be able to guarantee that
extra config values aren't being passed through the codebase to
dependencies outside of our control.