fix: prefer type-narrowing by elimination
We have a pattern in a few places where we perform type narrowing by
asserting the converse of the thing we actually care about. Take the
following:
```python
a: list[int] | None = [1, 2, 3]

assert isinstance(a, list)

assert a is not None

for item in a:
  print(item)
```

The first narrowing strategy is counterintuitively brittle. The thing
that we ultimately care about is that `a` is iterable. Switching `a` out
for any iterable type would continue to work. The other thing we care
about is that `a is not None` --- again any iterable would be fine here.

Ultimately, asserting a very specific type will typically lead to
rewritting chains of assertions whenever a type changes to an equivalent
but different type upstream. Instead narrowing by elimination will never
require rewritting (though it _could_ lead to unnecessary assertions,
but the type-checker will warn about that).