fix: remove pandas for temporary data structures
We were needing a way to organize some data about each tag in order to
move that data around from function->function. By sticking the data in
pandas, we fully erase any type information, information about the
column names, etc. from the type-checker and linter, making these tools
less useful to us.

To avoid needing to write a lot of small tests, it is better to rely on
static analysis tools when it is easy to do so --- and this change was
quite easy to do so.