chore: enable stricter pyright checking
In order to make pyright provide more useful instantaneous feedback on
code, opting into a few _slightly_ stricter rules.

**strictListInference, strictDictInference, strictSetInference**

all change the type inference behavior of container types. Specifically,
```python
a = [1, 'a', False]
```
could either be infered as `list[Any]` (the most narrow single type
that captures all of the elements) or `list[int | str | bool]`. The
second is more strict than the first, but allows avoid the propagation
of implicit `Any` throughout the code due to widened inference.

**enableReachabilityAnalysis**

makes the following an error:
```python
a: int | str

if isinstance(a, list):
  print('this is unreachable')
```

**reportInconsistentConstructor**

This is a more opinionated check that slightly limits python's OOP
capabilities --- however, it perfectly matches our current OOP practices
so should be a non-issue to adopt.

Specifically, this check ensures that the parent class `__init__` and
the child class `__init__` have compatible (e.g. matching) signatures.
This prevents the scenario where parents and children are intended to be
used interchangeably and so all children _must_ have the same signature
as the parent. This aligns with our "base classes are interfaces"
approach to OOP.

**reportMissingParameterType**

raises an error if a function signature does not have all parameters
either (a) inferred automatically e.g. `self` or (b) annotated. This
just helps ensure that we don't have any untyped functions which get
automatically inferred as `Unknown` which is equivalent to `Any` (i.e.
don't type-check this region of the code).

**reportUnnecessaryCast, reportUnnecessaryIsInstance**

both take advantage of type narrowing to ensure that we don't liter the
code with more type narrowing than is needed. For example, we currently
have _many_
```python
assert a is not None
```
due to allowing `None`s to propagated deeply throughout the code. As we
clean up the code, hopefully these `None`s will go away and we can rely
on the type-checker to flag asserts that are now made unnecessary.

**reportUnusedX**

raise an error if a class, expression, or function goes unused ---
helping us ensure that deadcode gets identified early and gets deleted
as we make changes.