refactor: remove reinforce algorithm
This algorithm operates quite differently from all other implemented
algorithms. Namely, it operates on full trajectories instead of
partial trajectories.

This isn't incompatible with the data pipeline --- the pipeline can be
asked for full trajectories instead of n-step trajectories (though
perhaps this is better facilitated with another transition creator);
however, the integration of the pipeline and Reinforce is currently
non-trivial.

Since Reinforce is only around as a historical artifact + something that
we _might_ want in the future as a baseline for research and experiments
it seems safe to kill for now and revive (and refactor) when/if the time
comes that we need it again.