chore: remove exploration module and rndnetwork
Currently the exploration is not wired up anywhere for any agent. It
appears as if the wiring code has decayed to the point that using the
exploration modules in their current state would require some
re-architecting.

In favor of heavily simplifying the codebase, removing this module for
now. If we decide to revisit, we can grep the commit history, though I
doubt this will need to happen (reimplementing RND isn't not too
challenging at all).