refactor: plumb ColumnDescription to agent
Goal is to get column name information and column ordering to the
agent. This should simplify the interface between the pipeline and the
agent, while allowing other column information to propagate more easily
(e.g. this will provide a mechanism to communicate endo/exo columns
without needing to rely on magical orderings).

Small risk factor: this couples the agent api to the pipeline api. This
is an acceptable risk while we wait for some tooling to bind column
names to Tensors being sent to the agent.