feat: perform sanity checking on interaction state
When data moves through the pipeline, the pipeline will filter out
nan values in the state _in the transition creator_. However, for
capturing a state for interactions, the capture_state hook happens
before the transition creator (which is good) which means that nan
filtering won't happen (which isn't good).

It should be the responsibility of the interaction object to ensure
a proper state is available and handed to the agent.

This change introduces several sanity checks and conditions for a state
to be considered valid before being handed over to the agent. This
includes checking for `NaN`s and checking for staleness of the data.