feat: add yellow, red, and black zones for tags
BREAKING CHANGE: this removes the concept of `bounds` from the tag
config, instead renaming these to `operating_range` for consistency.

Semantically, the operating_range (colloquially "black zone") specifies
a physically implausible bound. Values outside of this range are simply
incorrect. An example is the delta feature of the monotonically
increasing "total power consumption" tag in Epcor Solar. This must be
strictly non-negative, however sometimes it inexplicably becomes
negative. These occurrences imply something fundamentally incorrect in
the data stream. Another example: efficiency > 100% (also observed in
Epcor Scrubber).

The "red zone" specifies a range that the agent should actively try to
avoid. This is **not** a hard constraint, rather a strong preference.
An example includes a maximum acceptable pump speed.

The "yellow zone" specifies a warning zone that the agent should
typically avoid when possible. It can be inferred from a red zone.