refactor: migrate ensemble buffer to discrete-dists
This has a **massive** RAM impact. Previously, we were storing `ensemble`
number of replay buffers, each with a 50% chance of storing a piece of
data. Since `ensembles=10` by default, this implies 5x more RAM being
used than is strictly necessary.

This change instead makes the separation of elements part of the sampler
instead of part of the storage. Data is all stored in the same storage
backend (torch tensors in this case), with masked samplers handling
the bifurcation of the data into different buffers.