perf: set default low watermark to 1 for metrics
A low watermark of 1 means that the metrics writer tries writing
immediately when it receives the first hunk of data. While that write
process is occurring, the buffer will start to accumulate more data in a
secondary buffer, then will immediately flush that buffer once the first
write concludes.

In this way, the maximum latency of a given write is 2x the typical
latency of a python->tsdb write operation. There is never any waiting.