feat: add computed and sanitizer config utilities
The computed utility mutates an existing config key based on
intermediate config values found anywhere else in the config tree.
A trivial example is interpolation --- the computed value is a copy of
a value stored somewhere else on the config tree.

The sanitizer is a post-processing sanity checker. It can raise
arbitrary exceptions and asserts, which will get picked up by pydantic
and turned into nice error messages.

Both of these hook into the second pass of the config parser. This means
they can access values based on discriminated union resolutions. However
there is no guarantee on parsing or execution order, so computeds should
not rely on the outcome of other computeds.