feat: remove env.obs_period interpolations throughout configs
The various timing controls are now all contained within the interaction
which is a more natural home for them. All related interpolations are
now computeds and several needless interpolations (such as in the
pipline) have been removed to heavily simplify the total config tree.

BREAKING CHANGE: a result of these simplifications is that configs
that manipulated the timing controls within the env will no longer
be valid.