fix: move db infra configs into dedicated top-level config
This adds a new top-level config to specify our TSDB infrastructure
dependency in one place in the config. All other DBs in our config tree
will derive their configs from this.

NOTE: it is still possible to override _any_ of the defaults on a per-db
basis. That is, if we wanted to move our metrics table to a totally new
database, that would still possible with this change.

That is, our configs remain exactly as expressive now as before.