refactor: remove optional bootstrap_reduct arg
It isn't clear what the difference between the policy reduct and the
bootstrap reduct were. By default, these were identical and no configs
overrode that value. On top of this, there was an optional argument
to switch between them on a per-call basis --- but again, because both
were doing identical things, this was meaningless.

Even more confusing, several places in the code had (seemingly
arbitrary) different values of the optional arg, suggesting it must have
done _something_ at _some point_. All of this removed for conceptual
clarity.