style: annotate feature flags with creation date
Commonly, after a feature flag has been added for some amount of time,
teams will re-evaluate whether that feature still needs to be guarded.

Adding the time that the feature flag was added lets us know how long it
has been in a default "off" state. We should update that to the date
that it is defaulted to "on". After some amount of time, we should
remove the feature flag and simplify any conditional code depending on
that flag.