test: do not use uv run for highly parallel bsuite tests
`uv run` will first perform a `uv sync` before spinning up the job.
Typically, this is exactly what we want, however when we are doing
massively parallel jobs, this leads to a high probability of state
thrashing causing several jobs to fail.

Since we know the bsuite will forever be increasingly parallel, we
should assume the test environment is already configured and up to date.