fix: use deployment env/interaction by default
In general, our default values should be those used by the production
system--simulation and testing are the special case, not the other way
around. From this perspective, the environment and interaction should
be targeting their respective deployment versions.

This is causing schema validation errors to be meaningless when the
interaction or environment are unspecified. By providing a default, we
can have much stronger validation (essentially we can make more
assumptions and so get more value out of pydantic).