fix: wrap computeds and post-processes with ValueError handler
Pydantic only tries to capture `ValueError` exceptions when doing
custom validations. If any other type of exception is thrown, then
pydantic forwards the exception all the way through. This makes it
difficult to track down where the actual error is occurring.

By capturing arbitrary errors and transforming their type to
`ValueError`, we can get pydantic to produce a path and error message
for us. We will still hold on to the original stacktrace and dump that
to console to avoid masking real bugs.