fix: attempt to filter disc union values from error path
This is a known bug in pydantic:
https://github.com/pydantic/pydantic/issues/7261#issuecomment-1696927761

There are some smarter strategies that involve decorating disc unions
with some special error handler. This seems worth pursuing "correctly".
For now, the hacky solution gives us enough to not block UI work while
we put together a more robust solution on the backend.

Will create a new jira task to track the robust implementation.