feat: add standalone rejection sampling utility
This utility uses the oversampling strategy, where it asks for a
complete sample of fixed size on each iteration, then keeps a growing
list of valid samples.

In the happy path where we get the right number of samples on the first
iteration this is very slightly less efficient. In the unhappy path,
this tends to require far fewer iterations. Since we can parallelize
interior to an iteration, but not across, then this tends to be
substantially faster.