feat(PROD-375): replace linesearch with adam
Using an optimizer like adam with a small stepsize and many updates is
typically a much better gradient approximation than using linesearch
with a potentially large stepsize.

To keep the computational cost in check and to try to limit the degree
of overfitting, this change also incorporates an early stopping
mechanism. This mechanism is a bit naive in that it uses in-sample loss
as a stopping criteria (instead of out-of-sample data) and a very short
time horizon moving average over loss values. This, however, is likely
good enough considering the very small number of allowed updates on
each step.