refactor: use dependency injection to mock event_bus
This pattern separates out interfaces from implementations, allowing us
to more easily mock dependencies in unit tests. Specifically, we can now
easily mock the event bus for the zone discourager tests.

This does have the downside of duplicating interface code in a few
places, which increases the total amount of boilerplate. In this
particular case, it is worth it to avoid needing to spin up a network
connection in the tests. For other cases, it could cause more code
build-up than is helpful.