build: lock pyc_wheel version
There was a version bump from 1.3.0 -> 1.3.1 last night that changed
the way wheel files are generated, breaking our build. Ideally, we'd
have these types of build dependencies specified in our
pyproject.toml to make these less brittle, but for now manually locking
the version to unblock PRs in the pipeline.