fix(PROD-500): remove concept of base config path
This comes from seeking feature-parity with Hydra. Hydra is long gone,
and we are not using this feature meaningfully. Instead, now, we define
default config yamls using a relative path from the parent config.

In all current uses, this changes nothing --- the parent config and the
default config are in the same base directory.