refactor: remove collector from actor
The collector object is useful for research, but is a major risk for
production code. In particular, by default it collects metrics into
an in-memory sqlite database, which can grow indefinitely.

The current collection strategy is also very inefficient. Aggregating
metrics in memory, then pumping them to the collector in the research
code improves performance by nearly 50%.