feat(lib_agent): change actor stddev parameterization
By using a clip operation to control stochasticity of the actor,
we lose a lot of gradient information at the boundaries. We also can
easily get into an absorbing state where the actor stddev is very large
and we have no ability to change it do to zero gradients.

Instead of a one-tailed exponential with clipping, this change uses a
two-tailed exponential (aka a sigmoid) from (0, 0.1) to limit the
range of the stddev. In this case, we can only approach each end of the
range asymptotically and so will never experience true zero gradients.

The upper limit for stochasticity is somewhat arbitrary, but our
exploration strategy should not include a full uniform distribution over
actions. The wiggles should be localized. This **does not** mean that we
can't learn a bang-bang policy --- it just means we won't accidentally
be bang-bang due to learning a uniform distribution.

The claim is: we **never** want a uniform policy.