refactor(coreio): consolidate node registration logic
This lightly changes behavior by calling `ensure_connected` before
each registration, instead of calling once at the start of the loop.
This is a bit more costly since it makes more async calls to check
the health of the connection, however it adds a layer of safety. Since
this is not a hot path, this performance cost is acceptable.