docs(PROD-613): Add MVP diagram for backend
This diagram contains some temporal context within it, which is
generally a bad idea for documentation (has a clear expiration date).
Long-term intention is not to maintain the temporal context and will
remove after the planning phase is over in the next couple of weeks.

Useful to have this context for now as a communication tool.