refactor(PROD-341): switch to generic seasonal tag type
Seasonal tags had a different type for each seasonal tag. This was too
specific for the TagType. Instead, this commits switches to a more
broad categorization of "seasonal" tag type. Now the specific seasonal
tag is uniquely determined by the name.

NOTE: This is **not** a breaking change. Previously, the user had to
specify both the name and the tag type to be exactly the same. If the
user chose a different tag name, the code would break.

In addition to the above, the seasonal tags had a dummy extra config
schema that was 100% built with computed values. It was not actually
possible to modify these config values, instead they served only as a
cache. Instead, this can be implemented with a stateful stage.