fix(PROD-341): prefer immutable/covariant input type annotations
A function with a `list[T]` input type is invariant because we cannot
guarantee that the function won't mutate the list with locally valid
items that do not meet the outer scope's type guarantees. For example,
```python
class A:
    a: str

class B(A):
    b: int

def foo(x: list[A]):
    x.append(A())

list_of_b: list[B] = []
foo(list_of_b)  # this is a valid call

for item in list_of_b:
    print(item.b) # this will fail at runtime
```

By using immutable type annotations, we can ensure that the input type
is covariant, meaning that we can safely pass a list of a subclass type
to a function that expects a list of a superclass type. In this
scenario, the function `foo` above becomes invalid (an immutable
`Sequence` does not have an `append` method).