refactor(PROD-341): type guard against tag_type
Throughout the codebase, we have several methods that assume (or filter)
to operate on only a specific type of tag. This commit starts to
introduce type guards to these methods to allow correct type inference
once we've split `tag_type` into different subclasses.

In a few places, we do this type guarding with new `assert` statements.
This is **not** the final solution. Once we have split things into their
own types, we will be able to use smarter filters to put the impetus on
the caller code and leverage the type-checker to do the work for us.