fix(corerl): unify return of metrics read calls
These methods are not actually being consumed by anyone other than the
tests. We might consider nuking them as a result. But for now, since
there are no consumers, we can freely change the api. APIs that change
shape and structure as a result of config flags are devilishly hard to
manage, so this unifies the output to not change shape based on a flag.