feat(PROD-1112): add event bus client to corerl
This largely duplicates the event bus used p2p between corerl and
coreio, however it is designed for direct use with coredinator's global
event bus.

Intent is to deprecate the p2p connection and remove that code. In this
way, prefer **not** to share code between the two event bus clients to
facilitate that transition and reduce risk of cross-impacting changes.